% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOP.R
\name{BOP}
\alias{BOP}
\title{BOP}
\usage{
BOP(bar)
}
\arguments{
\item{bar}{- bars with which, to be evaluated}
}
\description{
This is balance of power. Basically, it shows where the price closed with respect to the length of the candle
}
\author{
Siva Sunku
}
\keyword{bhavcopy}
