% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.loadOneMinData.R
\name{data.getOneMinQuotes}
\alias{data.getOneMinQuotes}
\title{data.getOneMinQuotes}
\usage{
data.getOneMinQuotes(objectName, stock = NULL, fut = NULL,
  trim = "T23:55")
}
\arguments{
\item{objectName}{- objectname from which quotes to be read. This is generally from data.makeQuotesFromGDrive.}

\item{stock}{- Stock name to be filtered. Default ALL stocks}

\item{fut}{- default NULL. If fut prices are needed F1/F2 to be passed. currently only F1/F2 are available}

\item{trim}{- If True, only prices till 15:30 hours are returned. Default NULL.}
}
\description{
data.getOneMinQuotes
}
\details{
This gets the quotes of a particular stock from the big one min data object. This returns the object as xts for that particular stock.
}
