% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.price.hit.R
\name{is.price.hit}
\alias{is.price.hit}
\title{is.price.hit}
\usage{
is.price.hit(price, bars, closeLag = FALSE)
}
\arguments{
\item{price}{- price, that needs to be evaluated}

\item{bars}{- bars with which, to be evaluated}

\item{closeLag}{- if only close of previous bar to be evaluated. In this, only latest candle row is considered}
}
\description{
This will return if the given price is hit by the candlebar.
}
\note{

}
\author{
Siva Sunku
}
\keyword{price}
