% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.candle.R
\name{make.candle}
\alias{make.candle}
\title{make.candle}
\usage{
make.candle(bars)
}
\arguments{
\item{bars}{- bars with which, to be evaluated}
}
\description{
This will return a single bar of all the candles that were given
}
\note{

}
\author{
Siva Sunku
}
\keyword{make.candle}
